REFACTOR-#2739: io tests refactoring
